🎁 upd: shallow clone from latest tag/commit recorded in lockfile
(To obviously save time.)

The reason 1 day is subtracted from the recorded date is that sometimes
the shallow clone (using `--shallow-since`) doesn't work and does a full
clone instead. However, subtracting an entire day from the date seems to
consistently be able to shallow clone, with minimal extra cloned data.